fix(build): silence vite import warnings
