+++ docs/issues/vite-mixed-import-warnings/tasks.md
@@ -1,0 +1,7 @@
+# Vite Mixed Import Warnings Tasks
+
+- [x] Add SDD notes for the warning cleanup.
+- [x] Extract MIME detection helpers away from the adapter registry.
+- [x] Replace ineffective dynamic imports for legacy import and presenter singleton access.
+- [x] Verify no warned dynamic imports remain.
+- [x] Run required quality checks.
