+++ src/main/presenter/filePresenter/BaseFileAdapter.ts
@@ -2,6 +2,7 @@ import * as fs from 'fs'
 import * as crypto from 'crypto'
 import { FileMetaData } from '@shared/presenter'
 import path from 'path'
+import { detectMimeType } from './mimeDetection'
 
 export abstract class BaseFileAdapter {
   filePath: string
@@ -36,7 +37,6 @@ export abstract class BaseFileAdapter {
   }
 
   protected async preprocessFile(): Promise<void> {
-    const { detectMimeType } = await import('./mime')
     this.mimeType = await detectMimeType(this.filePath)
   }
